pyproject.toml
[project]
name = "mocklimit"
version = "0.1.0"
description = "Configurable mock API server with realistic rate limiting for testing"
readme = "README.md"
authors = [
    { name = "Stanislav Kosorin", email = "stanokosorin4@gmail.com" }
]
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
keywords = ["mock", "rate-limiting", "api", "testing", "openapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "fastapi>=0.135.1",
    "jsonref>=1.1.0",
    "loguru>=0.7.3",
    "pydantic>=2.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.42.0",
]

[project.urls]
Repository = "https://github.com/stano45/mocklimit"
Documentation = "https://github.com/stano45/mocklimit#readme"
Issues = "https://github.com/stano45/mocklimit/issues"

[project.scripts]
mocklimit = "mocklimit.main:main"

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "basedpyright>=1.38.3",
    "httpx>=0.28.1",
    "openai>=1.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]

[tool.ruff]
src = ["src"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D213"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "PLR2004"]

[tool.basedpyright]
typeCheckingMode = "strict"

[tool.pytest.ini_options]
asyncio_mode = "auto"
